requirements.txt
requests~=2.31.0
tqdm~=4.65.0
asyncio~=3.4.3
pandas~=2.0.3
tyro~=0.5.4
apscheduler~=3.10.1
numpy~=1.25.1
scipy~=1.11.1
jieba~=0.42.1
wordcloud~=1.9.2
matplotlib~=3.7.2
httpx~=0.24.1
setuptools~=68.0.0
bilibili-api-python~=15.5.1
opencv-python~=4.8.0.74